moved rdfmessageconverter to datastore
