+++ applications/search-api/pom.xml
@@ -45,6 +45,12 @@
             <groupId>org.elasticsearch</groupId>
             <artifactId>elasticsearch</artifactId>
         </dependency>
+
+        <dependency>
+            <groupId>no.dcat</groupId>
+            <artifactId>datastore</artifactId>
+            <version>0.2.4-SNAPSHOT</version>
+        </dependency>
     </dependencies>
 
     <!-- Nødvendig for å brukes på openshift-->
